{fuse, sched/wait}: Replicate the behavior of TASK_FREEZABLE on downstream
msm-5.4 lacks TASK_FREEZABLE which was introduced with freezer rewrite in
upstream [1]. That's a huge backport that's not worth for this small change.
So, instead, replicate how TASK_FREEZABLE works in upstream and adapt
for our kernel.

Now wait_event_freeze_state_exclusive() is only exclusive for freezable tasks
in fuse, not generic anymore. We cannot add freezer.h in wait.h because that
causes dependency issues. Instead move everything to fs/fuse/dev.c since only
fuse will be using this anyway.

[1] https://github.com/torvalds/linux/commit/f5d39b020809146

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>